refactor: move shared accessor helpers into utils
